+++ src/main/java/com/example/RealMatch/brand/application/service/BrandService.java
@@ -6,8 +6,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,11 +16,8 @@ import com.example.RealMatch.brand.domain.repository.BrandAvailableSponsorReposi
 import com.example.RealMatch.brand.domain.repository.BrandCategoryViewRepository;
 import com.example.RealMatch.brand.domain.repository.BrandLikeRepository;
 import com.example.RealMatch.brand.domain.repository.BrandRepository;
-import com.example.RealMatch.brand.exception.BrandErrorCode;
 import com.example.RealMatch.brand.presentation.dto.response.ActionDto;
 import com.example.RealMatch.brand.presentation.dto.response.BeautyFilterDto;
-import com.example.RealMatch.brand.presentation.dto.response.BrandCampaignResponseDto;
-import com.example.RealMatch.brand.presentation.dto.response.BrandCampaignSliceResponse;
 import com.example.RealMatch.brand.presentation.dto.response.BrandDetailResponseDto;
 import com.example.RealMatch.brand.presentation.dto.response.BrandFilterResponseDto;
 import com.example.RealMatch.brand.presentation.dto.response.SponsorInfoDto;
@@ -31,7 +26,6 @@ import com.example.RealMatch.brand.presentation.dto.response.SponsorProductDetai
 import com.example.RealMatch.brand.presentation.dto.response.SponsorProductListResponseDto;
 import com.example.RealMatch.campaign.domain.entity.Campaign;
 import com.example.RealMatch.campaign.domain.repository.CampaignRepository;
-import com.example.RealMatch.global.exception.CustomException;
 import com.example.RealMatch.global.presentation.advice.ResourceNotFoundException;
 import com.example.RealMatch.tag.domain.entity.BrandTag;
 import com.example.RealMatch.tag.domain.enums.TagType;
@@ -247,32 +241,5 @@ public class BrandService {
         return products.stream()
                 .map(SponsorProductListResponseDto::from)
                 .collect(Collectors.toList());
-    }
-
-    // 브랜드의 캠페인 리스트 조회
-    @Transactional(readOnly = true)
-    public BrandCampaignSliceResponse getBrandCampaigns(Long brandId, Long cursor, int size) {
-        brandRepository.findById(brandId)
-                .orElseThrow(() -> new CustomException(BrandErrorCode.BRAND_NOT_FOUND));
-
-        Pageable pageable = PageRequest.of(0, size + 1);
-        List<Campaign> campaigns = campaignRepository.findBrandCampaignsWithCursor(brandId, cursor, pageable);
-
-        boolean hasNext = campaigns.size() > size;
-        if (hasNext) {
-            campaigns = campaigns.subList(0, size);
-        }
-
-        LocalDateTime now = LocalDateTime.now();
-        List<BrandCampaignResponseDto> items = campaigns.stream()
-                .map(c -> new BrandCampaignResponseDto(
-                        c.getId(),
-                        c.getTitle(),
-                        c.getRecruitStartDate().toLocalDate(),
-                        c.getRecruitEndDate().toLocalDate(),
-                        c.getCampaignRecrutingStatus(now)
-                ))
-                .toList();
-        return new BrandCampaignSliceResponse(items, hasNext);
     }
 }
