+++ src/main/java/com/example/RealMatch/brand/presentation/controller/BrandCampaignController.java
@@ -1,0 +1,56 @@
+package com.example.RealMatch.brand.presentation.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.RealMatch.brand.application.service.BrandCampaignService;
+import com.example.RealMatch.brand.presentation.dto.response.BrandCampaignSliceResponse;
+import com.example.RealMatch.global.presentation.CustomResponse;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+
+@Tag(name = "Brand-Campain", description = "브랜드의 캠페인 API")
+@RestController
+@RequestMapping("/api/v1/brands")
+@RequiredArgsConstructor
+public class BrandCampaignController {
+    private final BrandCampaignService brandCampaignService;
+
+
+    @Operation(
+            summary = "브랜드의 캠페인 내역 조회 API by 박지영",
+            description = """
+                    브랜드의 캠페인을 조회합니다.   
+                    
+                    마지막으로 조회된 캠페인의 id를 cursor로 사용하는 페이징 방식입니다.   
+                    - 최초 조회 시 cursor 없이 요청합니다.     
+                    - 이후 조회 시 응답으로 받은 nextCursor 값을 cursor로 전달합니다.    
+                    - 정렬 조건 : UPCOMING /RECRUITING → 모집 시작 날짜 기준, CLOSED → 모집 끝나는 날짜 기준   
+                    
+                    캠페인 모집 상태는 다음 중 하나로 응답됩니다.   
+                    - UPCOMING   : 모집 예정   
+                    - RECRUITING : 모집 중   
+                    - CLOSED     : 완료 (모집 마감) 
+                    
+                    캠페인 모집 상태를 파라미터로 받을 수 있습니다. 현재 진행중인 브랜드의 캠페인은 RECRUITING 파리미터를 사용해주세요.
+                    ex. /api/v1/brands/{brandId}/campaigns?size=5&status="RECRUITING"
+                    """
+    )
+    @GetMapping("/{brandId}/campaigns")
+    public CustomResponse<BrandCampaignSliceResponse> getBrandCampaigns(
+            @Parameter(description = "브랜드 ID", example = "1")
+            @PathVariable Long brandId,
+            @RequestParam(required = false) Long cursor,
+            @RequestParam(defaultValue = "5") Integer size
+    ) {
+        BrandCampaignSliceResponse response = brandCampaignService.getBrandCampaigns(brandId, cursor, size);
+        return CustomResponse.ok(response);
+    }
+
+}
