+++ src/main/java/com/example/RealMatch/brand/presentation/swagger/BrandSwagger.java
@@ -2,11 +2,8 @@ package com.example.RealMatch.brand.presentation.swagger;
 
 import java.util.List;
 
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
 
-import com.example.RealMatch.brand.presentation.dto.response.BrandCampaignSliceResponse;
 import com.example.RealMatch.brand.presentation.dto.response.BrandDetailResponseDto;
 import com.example.RealMatch.brand.presentation.dto.response.BrandFilterResponseDto;
 import com.example.RealMatch.brand.presentation.dto.response.BrandLikeResponseDto;
@@ -67,29 +64,5 @@ public interface BrandSwagger {
     @Operation(summary = "브랜드 협찬 가능 제품 리스트 조회", description = "특정 브랜드의 협찬 가능 제품 목록을 조회합니다.")
     CustomResponse<List<SponsorProductListResponseDto>> getSponsorProducts(
             @Parameter(description = "브랜드 ID", required = true) @PathVariable Long brandId
-    );
-
-    @Operation(
-            summary = "브랜드의 캠페인 조회 API by 박지영",
-            description = """
-                    브랜드의 캠페인을 조회합니다.   
-                    
-                    마지막으로 조회된 캠페인의 id를 cursor로 사용하는 페이징 방식입니다.   
-                    - 최초 조회 시 cursor 없이 요청합니다.     
-                    - 이후 조회 시 응답으로 받은 nextCursor 값을 cursor로 전달합니다.    
-                    - 정렬 조건 : UPCOMING /RECRUITING → 모집 시작 날짜 기준, CLOSED → 모집 끝나는 날짜 기준   
-                    
-                    캠페인 모집 상태는 다음 중 하나로 응답됩니다.   
-                    - UPCOMING   : 모집 예정   
-                    - RECRUITING : 모집 중   
-                    - CLOSED     : 완료 (모집 마감) 
-                    """
-    )
-    @GetMapping("/{brandId}/campaigns")
-    CustomResponse<BrandCampaignSliceResponse> getBrandCampaigns(
-            @Parameter(description = "브랜드 ID", example = "1")
-            @PathVariable Long brandId,
-            @RequestParam(required = false) Long cursor,
-            @RequestParam(defaultValue = "10") Integer size
     );
 }
