+++ src/main/java/com/example/RealMatch/campaign/domain/enums/CampaignRecruitingStatus.java
@@ -1,6 +1,6 @@
 package com.example.RealMatch.campaign.domain.enums;
 
-public enum CampaignRecrutingStatus {
+public enum CampaignRecruitingStatus {
     UPCOMING,   // 모집 예정
     RECRUITING, // 모집 중
     CLOSED      // 모집 마감
